Windows compat. Create quick demo gif of v0

+++ code/shell.h
@@ -47,5 +47,5 @@ private:
 
 	int m_cursorPos;
 	std::string m_stashedInput;
-	int m_timer;
+	float m_timer;
 };
